requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Web framework
flask==3.0.0
jinja2==3.1.2

# Data processing
pandas==2.1.4
numpy==1.24.3
pyspark==3.5.0

# ML and NLP
transformers==4.36.0
sentence-transformers==2.2.2
torch==2.1.0
scikit-learn==1.3.2

# Vector databases
chromadb==0.4.18
pinecone-client==2.2.4

# LLM providers
groq==0.4.1
openai==1.3.7

# MLflow and tracking
mlflow==2.8.1

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# API utilities
slowapi==0.1.9
prometheus-client==0.19.0

# Document processing
unstructured[pdf]==0.11.6
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Logging and monitoring
structlog==23.2.0